### 2022-01-21 #4091 @kyleniemeyer
Hi @szwiep, I think the empty `orcid:` field under your name is causing the paper build issue. Can you either add your ORCID, or remove the field?
