### 2022-01-21 #4091 @kyleniemeyer
@szwiep ah, I'm sorry, it was actually the affiliation field for your coauthor that was causing the problem. (Or perhaps both things.)

Could you change `affiliation: 1, 2` to `affiliation: "1, 2"` (with the quotation marks)?
