### 2022-01-21 a8f800c
Update paper.md
Add quotations to affiliation field
