### 2022-01-21 #4091 @kyleniemeyer
@szwiep this just won't quite resolve! I notice one issue but miss the others, apparently...

The sub-fields under your coauthor's name should not have dashes, since that is used to mark new entries in the `authors` list. So, can you replace the second author block with
```yaml
  - name: Shawn M. Chartrand^[co-first author]
    orcid: 0000-0002-9309-1137
    affiliation: "1, 2"
```
