### 2022-01-21 #4091 @szwiep
@kyleniemeyer Thanks for helping sort this out. I'll update the dashes. One potential problem: would the second affiliation need an `index: 2` sub-field? Right now it doesn't.
