### 2022-01-21 e802e2c
Update paper.md
Remove dashes from Author sub-fields
