### 2022-01-21 #4091 @kyleniemeyer
@szwiep no, combining two affiliations within quotes like that is the correct approach. (See the example paper: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography)
