### 2022-01-21 27c6b3b
Update paper.md
Add index to second affiliation
