### 2022-01-21 #4091 @szwiep
@kyleniemeyer sorry, I meant in the affiliations section. Looks like we we're missing the index:

```{markdown}
  - name: Department of Earth Sciences, Simon Fraser University
   index: 2
```
Just committed those (dashes and index) changes.
