### 2022-01-21 #4091 @kbarnhart
@kyleniemeyer I can handle this. I'll go ahead and assign myself.

@szwiep I'll be able to start handling this actively in the middle of next week. At that point I'll do a standard pre-review check of the submission, depending on the results of that I may provide some feedback, and then will start to look for reviewers. Please let me know if you have questions at any point either here or at krbarnhart@usgs.gov
