+++ paper/paper.bib
@@ -54,7 +54,8 @@
 @phdthesis{Chartrand:2017, 
 	series={Electronic Theses and Dissertations (ETDs) 2008+}, 
 	title={Pool-riffle dynamics in mountain streams : implications for maintenance, formation and equilibrium}, 
-	url={https://open.library.ubc.ca/collections/ubctheses/24/items/1.0349138}, DOI={http://dx.doi.org/10.14288/1.0349138}, 
+	url={https://open.library.ubc.ca/collections/ubctheses/24/items/1.0349138}, 
+	doi={10.14288/1.0349138}, 
 	school={University of British Columbia}, 
 	author={Chartrand, Shawn M.}, 
 	year={2017}, 
@@ -144,6 +145,7 @@
 	address = {Reston, VA}, 
 	title = {Transport of gravel and sediment mixtures}, 
 	booktitle = {Sedimentation Engineering: Theory, Measurements, Modeling and Practice (ASCE Manuals and Reports on Engineering Practice No. 110)}, 
+	doi = {10.1061/9780784408148.ch03},
 	publisher = {ASCE}, 
 	author = {Parker, Gary}, 
 	editor = {Garcia, MH}, 
@@ -172,7 +174,7 @@
 	number = {4},
 	pages = {1155-1202},
 	keywords = {sediment transport, geomorphology, turbulence, complex systems, models, management},
-	doi = {https://doi.org/10.1002/2014RG000474},
+	doi = {10.1002/2014RG000474},
 	year = {2015}
 }
 
