### 2022-01-23 #4091 @smchartrand
Just a note that I have addressed the doi issues raised above by @whedon during the pdf build process.
