### 2022-01-27 #4091 @smchartrand
@kbarnhart - I am following up on the above request for suggested reviewers. After looking through the JOSS reviewer list, here are some suggestions (I searched for keyword "geomorphology" and started at the bottom of the list and made it halfway):

1. marda-science and dbuscombe-usgs
2. ldominguezruben
3. katmratliff
4. pwernette
5. chrisleaman
6. kellykochanski
