### 2022-01-27 #4091 @kbarnhart
@szwiep @smchartrand Thanks for this submission, I think it is a great contribution and I look forward to seeing it through the JOSS review process. I've gone through my standard pre-review evaluation of this submission and I have identified some issues that I think would be beneficial to address before I find reviewers and start review. Most of these comments are directly related to the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) and [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html). These are issues that would come up very quickly during review and so I find that it is best to address them before review starts.

* **Documentation**
   * **Missing API documentation.** I see most functions are well documented, but I couldn't find API documentation linked from the readme. Based on the readme, I know that the main intent is for the model to be run by specifying a .yaml file, but the extensibility of the submission relies on the core functions themselves. Thus, I’d expect the API level documentation to be exposed.
    * **Examples.** Other than the readme, I didn't find any examples. Examples don't need to be extensive, but some more info to help a user understand how to use the software. An example would also provide the possibility to show an example .yaml input file.
    * **Theory:** Somewhere in the documentation I'd recommend a bit more theory background connecting the software options to the theory of rarified sediment transport (e.g., why a lognormal distribution, how do you handle when particles land on one another).
    * **More details regarding the output file.** Some more information about the output file would benefit a user. For example: how are the final metrics calculated (including the subregion definition), what is stored in `np.array(f["iteration_i"]["model"])`? Particle ages, locations, depths, etc?
    * **Needs additional installation architecture/instructions** A user might want to have their project directories separate from this package (e.g., alternative file structures than what is shown in the readme, including not needing to duplicate the python code provided here). How might a user handle this? I suspect that it would be best solved by more formally packaging the contribution so that py_SBeLT could be installed and then run.py and plot_maker.py could be run from the command line using a [console_script](https://packaging.python.org/en/latest/guides/distributing-packages-using-setuptools/?highlight=console_scripts#console-scripts).
* **Tests:** I see that the submission has tests, but it is not clear how a user would run them.
* **Address reference formatting**. For example, disambiguation (D.J. vs David Jon Furbish) and missing citekeys, (e.g., lines 12).

I'm going to mark this submission as paused to give you whatever time you need to address the above comments. Please consider me a resource in addressing these comments and let me know if you have any questions. You can reach me either here or at krbarnhart@usgs.gov
