### 2022-02-01 #4091 @smchartrand
Quick question @kbarnhart. Your above comment regarding **Theory** - is it best to add the additional text in the readme or the JOSS paper? Thanks.
