### 2022-02-01 #4091 @kbarnhart
I'd recommend adding it to the ReadMe or another page in the documentation (JOSS in flexible in how the documentation is structured). That way the theory is more easily discoverable relative to the code.
