### 2022-02-01 4330a40
Update THEORY.md
