### 2022-02-01 f2bf77b
Update THEORY.md
