### 2022-02-01 b0f1f45
Update THEORY.md
