### 2022-02-01 4e21c5a
Update THEORY.md
