### 2022-02-06 fe76347
Update THEORY.md
