### 2022-02-07 28940c2
Update THEORY.md
