+++ THEORY.md
@@ -74,8 +74,13 @@ to generating quasi-random sets of travel distances.
 The selected probability density function is specified within the **'parameter.yaml'** file (see readme.md and paper.md), and the readme.md provides the 
 distribution function parameter value ranges tested to date. The gamma distribution, for example, can also provide modes displaced from zero, and with 
 distribution shapes that are skewed to longer lengths. Furthermore, physical experiments that in part motivated development of **`py_SBeLT`** report that particle 
-hop distances are well described by the Weibull distribution (Fathel et al., 2015). These additional pdfs can be easily incorporated into future model extensions 
-using the **'scipy.stats'** library of functions.
+hop distances are well described by the Weibull distribution (Fathel et al., 2015). Note however that **'py_SBeLT'** would likely produce isolated particle piles 
+if the shape parameter $k$ were constrained to values <1 (Fathel et al., 2015). These additional pdfs can be easily incorporated into future model extensions 
+using the **'scipy.stats'** library of functions with appropiate modifications to the **'parameter.yaml'** and **'run.py'** files.
+
+|![Image](/figures/weibull.png)
+|:--:| 
+| *Figure 3. Example Weibull pdf that could be used to set the number of entrainment events in **`py_SBeLT`**.* |
 
 ## References
 
