+++ sbelt/run.py
@@ -1,0 +1,289 @@
+import numpy as np
+import yaml
+import logging
+import logging.config
+from datetime import datetime
+from pathlib import Path 
+import h5py
+import time
+from tqdm import tqdm
+from jsonschema import validate, exceptions
+
+import sbelt.logic as logic
+import sys
+import os
+
+ITERATION_HEADER = ('Beginning iteration {iteration}...')
+ENTRAINMENT_HEADER = ('Entraining particles {event_particles}')
+
+def main():
+
+    # pid = os.getpid()
+    # run_id = datetime.now().strftime('%y%m-%d%H')
+    # # TODO: send to default config, or get user to edit config, or ask for path to config
+    # param_path = sys.argv[1]
+
+    logConf_path, log_path, schema_path, output_path = get_relative_paths()
+
+    #############################################################################
+    # Set up logging
+    #############################################################################
+    
+    configure_logging(run_id, logConf_path, log_path)
+    
+    #############################################################################
+    # Get and validate parameters
+    #############################################################################
+    
+    with open(schema_path, 'r') as s:
+        schema = yaml.safe_load(s.read())
+    with open(param_path, 'r') as p:
+        parameters = yaml.safe_load(p.read())
+    try:
+        validate(parameters, schema)
+    except exceptions.ValidationError as e:
+        print("Invalid configuration of param file at {param_path}. See the exception below:\n" )
+        raise e
+    if parameters['x_max'] % parameters['set_diam'] != 0:
+        print("Invalid configuration of param file at {param_path}: x_max must be divisible by set_diam.")
+        raise ValueError("x_max must be divisible by set_diam")
+    if parameters['x_max'] % parameters['num_subregions'] != 0:
+        print("Invalid configuration of param file at {param_path}: x_max must be divisible by num_subregions.")
+        raise ValueError("x_max must be divisible by num_subregions")
+
+    #############################################################################
+    #  Create model data and data structures
+    # TODO: Better names for d, h variables
+    #############################################################################
+
+    print(f'[{pid}] Building Bed and Model particle arrays...')
+    # Pre-compute d and h values for particle elevation placement
+        # see d and h here: https://math.stackexchange.com/questions/2293201/
+    d = np.divide(np.multiply(np.divide(parameters['set_diam'], 2), 
+                                        parameters['set_diam']), 
+                                        parameters['set_diam'])
+    h = np.sqrt(np.square(parameters['set_diam']) - np.square(d))
+    # Build the required structures for entrainment events
+    bed_particles, model_particles, model_supp, subregions = build_stream(parameters, h)
+
+    #############################################################################
+    #  Create entrainment data and data structures
+    #############################################################################
+
+    particle_age_array = np.ones(parameters['n_iterations'])*(-1)
+    particle_range_array = np.ones(parameters['n_iterations'])*(-1)
+    snapshot_counter = 0
+
+    #############################################################################
+
+    h5py_filename = f'{parameters["filename_prefix"]}-{run_id}.hdf5'
+    hdf5_path = f'{output_path}/{h5py_filename}'
+
+    with h5py.File(hdf5_path, "a") as f: 
+        
+        grp_p = f.create_group(f'params')
+        for key, value in parameters.items():
+            grp_p[key] = value
+
+        grp_iv = f.create_group(f'initial_values')
+        grp_iv.create_dataset('bed', data=bed_particles)
+        grp_iv.create_dataset('model', data=model_particles)
+
+        #############################################################################
+        #  Entrainment iterations
+        #############################################################################
+
+        print(f'[{pid}] Bed and Model particles built. Beginning entrainments...')
[…]
