+++ sbelt/plots/plotting.py
@@ -1,0 +1,189 @@
+"""
+All things plotting related.
+Most of this code created by Dr. Shawn Chartrand.
+"""
+import numpy as np
+import matplotlib
+from matplotlib import pyplot as plt
+from matplotlib.collections import PatchCollection
+from matplotlib.patches import Circle
+from scipy.optimize import curve_fit
+from scipy.special import factorial
+from PIL import Image
+
+def stream(iteration, bed_particles, model_particles, x_lim, y_lim, fp_out):
+    """ Plot the complete stream from 0,0 to x_lim and y_lim. Bed particles 
+    are plotted as light grey and model particles are platted in a colour
+    range dependant on their age. 
+    Allows for closer look at state of a subregion of the stream during simulation.
+    Also makes for fun gifs.
+
+    Keyowrd arguments:
+        iteration -- the iteration of the stream being plotted 
+        bed_particles -- array of all bed particles
+        model_particles -- array of all model particles
+        x_lim -- length of stream to plot
+        y_lim -- height of stream to plot
+        fp_out -- save location
+    """
+    plt.clf()
+    fig = plt.figure(1)
+    fig.set_size_inches(20, 6.5)
+    ax = fig.add_subplot(1, 1, 1, aspect='equal')
+    # NOTE: xlim and ylim modified for aspec ratio -- WIP
+    ax.set_xlim((-2, x_lim))
+    ax.set_ylim((0, y_lim))
+    
+    radius_array = np.asarray((bed_particles[:,1] / 2.0), dtype=float)
+    x_center = bed_particles[:,0]
+    y_center_bed = np.zeros(np.size(x_center))
+    plt.rcParams['image.cmap'] = 'gray'
+    ## This method of plotting circles comes from Stack Overflow questions\32444037
+    ## Note that the patches won't be added to the axes, instead a collection will.
+    patches = []
+    for x1, y1, r in zip(x_center, y_center_bed, radius_array):
+        circle = Circle((x1, y1), r)
+        patches.append(circle)
+    p = PatchCollection(patches, color="#BDBDBD", alpha=0.9, linewidths=(0, ))
+    ax.add_collection(p)
+    
+    x_center_m = model_particles[:,0]
+    y_center_m = model_particles[:,2]
+    patches1 = []
+    for x2, y2, r in zip(x_center_m, y_center_m, model_particles[:,1]/2):
+        circle = Circle((x2, y2), r)
+        patches1.append(circle)
+    p_m = PatchCollection(patches1, cmap=matplotlib.cm.RdGy, edgecolors='black')
+    p_m.set_array(model_particles[:,5])
+    ax.add_collection(p_m)
+
+    plt.colorbar(p_m,orientation='horizontal',fraction=0.046, pad=0.1,label='Particle Age (iterations since last hop)')
+    plt.title(f'Iteration {iteration}')
+
+    filename = f'iter{iteration}.png'
+    plots_path = fp_out + filename
+    plt.savefig(plots_path, format='png',)
+        
+    return
+
+def stream_gif(start, stop, dir):
+    """Plot gif of the streambed"""
+    in_filename = 'iter{i}.png'
+    out_filename = 'simulation_snapshot.gif'
+
+    fp_in = dir + in_filename
+    fp_out = dir + out_filename
+
+    step = 1
+    images = []
+
+    for i in range(start, stop, step):
+        im = Image.open(fp_in.format(i=i))
+        images.append(im)
+
[…]
