+++ model/logic.py
@@ -1,616 +1,0 @@
-import math
-import random
-import numpy as np
-
-
-import logging
-
-
-class Subregion():
-    """ Subregion class.
-    
-    Each instance of Subregion contains
-    the name, and the left and right 
-    boundaries of a subregion. 
-    
-    Name and boundaries are set during 
-    initualization and can be retrieved
-    afterwards using helper methods.
-    """
-    def __init__(self, name, left_boundary, right_boundary, iterations):
-        self.name = name
-        self.left_boundary = left_boundary
-        self.right_boundary = right_boundary
-        self.flux_list = np.zeros(iterations, dtype=np.int64)
-        
-    def leftBoundary(self):
-        return self.left_boundary
-    
-    def rightBoundary(self):
-        return self.right_boundary
-    
-    def getName(self):
-        return self.name
-
-    def incrementFlux(self, iteration):
-        self.flux_list[iteration] += 1
-    
-    def getFluxList(self):
-        return self.flux_list
-
-def get_event_particles(e_events, subregions, model_particles, level_limit, height_dependant=False):
-    """ Find and return list of particles to be entrained
-
-    Keyword arguments:
-    e_events -- number of events requested per subregion 
-    subregions -- array of Subregion objects
-    model_particles -- array of all model particles
-
-    Returns:
-    event_particles -- List of particles to be entrained
-
-    """
-    if e_events == 0:
-        e_events = 1 #???
-    
-    event_particles = []
-    for subregion in subregions:
-        # Take only particles in the boundaries of the current subregion
-        subregion_particles = model_particles[
-                (model_particles[:,0] >= subregion.leftBoundary())
-              & (model_particles[:,0] <= subregion.rightBoundary())]
-        # Take only particles that are in-stream (not ghost)
-        in_stream_particles = subregion_particles[
-                                                subregion_particles[:,0] != -1]
-        # Take only particles that are 'active' 
-        active_particles =  in_stream_particles[
-                                                in_stream_particles[:,4] != 0]
-
-        # Do not take any particles that have been selected for entrainment (i.e do not double select)
-        # This only happens when particles rest on the boundary. 
-        active_event, active_idx, event_idx = np.intersect1d(active_particles[:,3], event_particles, return_indices=True)
-        active_particles = np.delete(active_particles, active_idx, axis=0)
-
-        subregion_event_ids = []  
-        if height_dependant: # any particle at the level limit must be entrained
-            levels = elevation_list(subregion_particles[:,2], desc=False)
-            tip_particles = []
-            # find the tip particles -- these are the particles being entrained
-            if len(levels) == level_limit: 
-                tip_particles = active_particles[active_particles[:,2] == levels[level_limit-1]]
-            for particle in tip_particles:
-                subregion_event_ids.append(particle[3])
-                active_particles = active_particles[active_particles[:,2] != particle[2]]
-        # If there are not enough particles in the subregion to sample from, alter the sample size
-        if e_events > len(active_particles):
-            random_sample = random.sample(range(len(active_particles)), 
-                                        len(active_particles))
-        else: 
-            random_sample = random.sample(range(len(active_particles)), 
-                                        e_events)
-        # TODO: change so that we don't rely on loop index to grab particle
-        for index in random_sample:
-            subregion_event_ids.append(int(active_particles[index][3])  )
-        
-        ghost_particles = np.where(model_particles[:,0] == -1)[0]
-        for index in ghost_particles:
-            model_particles[index][0] = 0 
-            subregion_event_ids.append(index)
-        
-        if e_events != len(subregion_event_ids):
-            msg = (
[…]
