+++ requirements.txt
@@ -1,9 +1,8 @@
-matplotlib==3.3.4
-shortuuid==1.0.1
 h5py==2.9.0
 tqdm==4.56.0
-scipy==1.6.0
 jsonschema==3.0.2
-numpy==1.19.5
-Pillow==9.0.0
+matplotlib==3.3.4
+numpy==1.21.5
+scipy==1.6.0
+Pillow==9.0.1
 PyYAML==6.0
