### 2022-03-11 8316f83
Structural changes for packaging for library-like existence
