### 2022-03-12 1a975ac
Docstrings as far as the eye can see, API HTML docs, and restructuring
