+++ src/tests/test_utils.py
@@ -1,0 +1,29 @@
+"""
+A module for unit tests of the utils module
+
+Todo:
+    * Complete these unit tests. Not finished in version
+    1.0.0 b/c invalid (if validation fails) parameters will 
+    simply cause the model to fail silently at a later point.
+    A headache to resolve, but not impedeing functionality.
+"""
+import unittest
+from sbelt import utils
+
+class TestValidateArguments(unittest.TestCase):
+
+    def test_valid_set_of_arguments_passes(self):
+        print("Not implemented.")
+        return 
+    
+    def test_invalid_types_raise_valueError(self):
+        print("Not implemented.")
+        return 
+    
+    def test_invalid_ranges_raise_valueError(self):
+        print("Not implemented.")
+        return 
+    
+    def test_invalid_bed_subr_diam_raise_valueError(self):
+        print("Not implemented.")
+        return 
