+++ src/tests/test_logic.py
@@ -1,11 +1,16 @@
+"""
+A module for unit tests of the logic module
+
+Todo:
+    * automate the creation of particle arrays
+    so that it isn't harcoded in each test func
+"""
+
 import unittest
-import random
-import math
-from unittest.case import expectedFailure
 import numpy as np
 from unittest.mock import Mock
 
-from ..sbelt import logic
+from sbelt import logic
 
 ATTR_COUNT = 7 # Number of attributes associated with a Particle
 # For reference:
@@ -18,7 +23,21 @@ ATTR_COUNT = 7 # Number of attributes associated with a Particle
         # [6] = loop age counter
   
 class TestGetEventParticlesWithOneSubregion(unittest.TestCase):
+    """
+    Test that getting event particles with one Subregion 
+    returns a valid list of event particles.
+    
+    A 'valid list' will change depending on the function.
+    See function docstrings for more details.
 
+    Attributes:
+        test_length: the length of the bed
+        num_particles: the number of model particles
+        mock_sub_list: list of Mock-type subregions
+        entrainment_events: number of entrainment events to request
+            per subregion
+        level_limit: random int representing level limit
+    """
     def setUp(self):
         self.test_length = 10
         self.num_particles = 3
@@ -32,8 +51,10 @@ class TestGetEventParticlesWithOneSubregion(unittest.TestCase):
         self.entrainment_events = 3
         self.level_limit = np.random.randint(0, np.random.randint(2, 10))
 
-
     def test_all_active_returns_valid_list(self): 
+        """If there are N active particles in 1 subregion and N events requested
+        per subregion then a valid list will be a list of all particles.
+        """
         model_particles = np.zeros((self.num_particles, ATTR_COUNT))
         model_particles[:,3] = np.arange(self.num_particles) # unique ids
         model_particles[:,4] = np.ones(self.num_particles) # all active
@@ -58,9 +79,11 @@ class TestGetEventParticlesWithOneSubregion(unittest.TestCase):
         self.assertCountEqual(hp_list, model_particles[:,3])
         self.assertCountEqual(hp_list, list)
     
-    # TODO: Mock logging object and assert warning is logged
     def test_not_all_active_returns_list_of_2(self):
-
+        """If there are N particles in 1 subregion and N-1 are _active_, 
+        and if N events are requested per subregion then a valid list will be 
+        a list of the two active particles.
+        """
         mp_one_inactive = np.zeros((self.num_particles, ATTR_COUNT))
         mp_one_inactive[:,3] = np.arange(self.num_particles) 
         mp_one_inactive[0][4] = 1
@@ -76,9 +99,11 @@ class TestGetEventParticlesWithOneSubregion(unittest.TestCase):
         active_list = mp_one_inactive[mp_one_inactive[:,4] != 0]
         self.assertCountEqual(list, active_list[:,3])
 
-    # TODO: Mock logging object and assert warning is logged
     def test_none_active_returns_empty_list(self):
-        
+        """If there are N particles in 1 subregion and 0 are _active_ 
+        and if N events are requested per subregion, then a valid list will be 
+        an empty list.
+        """
         np_none_active = np.zeros((self.num_particles, ATTR_COUNT))
         np_none_active[:,3] = np.arange(self.num_particles) 
         np_none_active[:,0] = np.random.randint(self.test_length, size=self.num_particles)
@@ -92,7 +117,11 @@ class TestGetEventParticlesWithOneSubregion(unittest.TestCase):
     
 
     def test_all_ghost_particles_returns_ghost_particles(self):
-        
+        """If there are N particles in 1 subregion and all N particles
+        are 'ghost' particles (at -1), and if N particles are requested
+        per subregion, then a valid list will be a list of all the
+        ghost particles (all the particles).
+        """
[…]
