+++ requirements.txt
@@ -1,8 +1,6 @@
+scipy==1.6.0
+numpy==1.21.5
 h5py==2.9.0
 tqdm==4.56.0
-jsonschema==3.0.2
 matplotlib==3.3.4
-numpy==1.21.5
-scipy==1.6.0
 Pillow==9.0.1
-PyYAML==6.0
