+++ src/logs/conf.yaml
@@ -1,23 +1,0 @@
-version: 1
-formatters:
-  simple:
-    format: '%(levelname)s - %(message)s'
-handlers:
-  # console:
-  #   class: logging.StreamHandler
-  #   level: INFO
-  #   formatter: simple
-  #   stream: ext://sys.stdout
-  file:
-    class : logging.handlers.RotatingFileHandler
-    formatter: simple
-    filename: ''
-    backupCount: 3
-loggers:
-  sampleLogger:
-    level: WARN
-    handlers: [file]
-    propagate: no
-root:
-  level: WARN
-  handlers: [file]

[…]
