+++ src/sbelt/logic.py
@@ -1,3 +1,20 @@
+""" 
+This module contains all the logic required to execute
+sbelt simulations/runs. All functions/classes are designed 
+for internal use and may change without note.
+
+A primary purpose of this module is to manipulate n-7 NumPy 
+arrays which represent multiple stream particles. In these
+n-7 arrays, a _single_ particle is represented by a NumPy array 
+with 7 attributes::
+
+    [x_location, diam, y_location, UID, active state, age, loop age]
+
+In this general example, x_location and y_location define the centre point of 
+spherical particle whose diameter is defined by diam. See the project docs for
+more information on the other attributes. 
+"""
+
 import math
 import random
 import numpy as np
@@ -15,9 +32,9 @@ class Subregion():
     subregion maintains a NumPy list which is 
     used to record the number of model particles that 
     pass the downstream boundary in a given iteration. 
-    For example:
+    For example::
 
-    flux_list = [0,3,2,1]
+        flux_list = [0,3,2,1]
 
     Means that 0 crossings happened in the first iteration, 
     3 happened in the 2nd, and so on. The list has 
@@ -64,7 +81,7 @@ def get_event_particles(e_events, subregions, model_particles, level_limit, heig
 
     Will loop through each subregion and select n = e_events
     model particles (within a subregion boundaries) at random 
-    to be entrained.
+    to be entrained. No particle will be selected twice.
 
     Args:
         e_events: The number of events requested per subregion (int).
@@ -74,13 +91,12 @@ def get_event_particles(e_events, subregions, model_particles, level_limit, heig
 
     Returns:
         event_particles: A NumPy array of k uids representing the model particles
-            that have been selected for entrainment. For example:
+            that have been selected for entrainment. For example::
 
-            [2.0, 5.0, 25.0]
+                [2.0, 5.0, 25.0]
 
             Will represent that model particles with uids 2.0, 5.0 and
-            25.0 have been selected for entrainment. No particle uuid 
-            will appear twice in the array.
+            25.0 have been selected for entrainment. 
     """
     if e_events == 0:
         e_events = 1 #???
@@ -172,10 +188,10 @@ def build_streambed(bed_length, particle_diam):
     
     Returns:
         bed_particles: An m-7 NumPy array representing the stream's m 
-            bed particles. For example:
+            bed particles. For example::
 
-            [[x1, diam, y, uid1, active, age, loops], ... 
-                        ,[xM, diam, y, uidM, active, age, loops]]
+                [[x1, diam, y, uid1, active, age, loops], ... 
+                            ,[xM, diam, y, uidM, active, age, loops]]
             
             Where all bed particles share the same diam (diam1=...=diamM) and 
             y (y1=...=yM), all uids are negative, and to represent 'static-ness'
@@ -289,14 +305,12 @@ def find_supports(particle, model_particles, bed_particles):
 
     Provided a particle p at location x, this function 
     will search the stream for particles that p would
-    rest on, if being dropped at x. 
-    
-    More generally, supporting particles are those 
-    particles which directly hold up a particle. Supporting
+    rest on, if being dropped at x. More generally, supporting particles 
+    are those particles which directly hold up a particle. Supporting
     particles will always have a centre location that is 
     exactly a radius length away from p's centre.
 
-    Args:   
+    Args:
         particle: 1-7 NumPy array representing a model particle.
         model_particles: An n-7 NumPy array representing the stream's 
             n model particles.
@@ -345,16 +359,6 @@ def set_model_particles(bed_particles, available_vertices, particle_diam, pack_f
     Model particles are randomly placed at available vertex
     locations (x,y) across the bed. Location and initial attribute
     values are stored in the returned NumPy array. 
-
[…]
