### 2022-03-17 7e1a3c2
Update doc strings for proper pdoc3 parsing
