### 2022-03-18 0316a33
Update test command for CircleCI
