### 2022-03-18 4edeb35
Update plotting functions to work in package
