+++ src/sbelt/plots/plotting.py
@@ -9,14 +9,12 @@ from matplotlib.collections import PatchCollection
 from matplotlib.patches import Circle
 from scipy.optimize import curve_fit
 from scipy.special import factorial
-from PIL import Image
 
-def stream(iteration, bed_particles, model_particles, x_lim, y_lim, fp_out=None, out_name=None):
+def stream(iteration, bed_particles, model_particles, x_lim, y_lim, out_location=None, out_name=None):
     """ Plot the complete stream from 0,0 to x_lim and y_lim. Bed particles 
-    are plotted as light grey and model particles are platted in a colour
-    range dependant on their age. 
-    Allows for closer look at state of a subregion of the stream during simulation.
-    Also makes for fun gifs.
+    are plotted as light grey and model particles are plotted in a colour
+    range dependant on their age. Allows for closer look at state of a subregion of the 
+    stream during simulation. Also makes for fun gifs!
 
 
     Args:
@@ -25,12 +23,14 @@ def stream(iteration, bed_particles, model_particles, x_lim, y_lim, fp_out=None,
         model_particles: array of all model particles
         x_lim: length of stream to plot
         y_lim: height of stream to plot
-        fp_out: save location
-        out_name: filename (ignored if fp_out not set) 
+        out_location: save location
+        out_name: filename (ignored if out_location not set) 
     """
-    plt.clf()
+    if out_location is not None:
+        if out_name is None:
+            raise ValueError('The out_name argument must be set if saving file.')
     fig = plt.figure(1)
-    fig.set_size_inches(20, 6.5)
+    # fig.set_size_inches(20, 6.5)
     ax = fig.add_subplot(1, 1, 1, aspect='equal')
     # NOTE: xlim and ylim modified for aspec ratio -- WIP
     ax.set_xlim((-2, x_lim))
@@ -62,54 +62,33 @@ def stream(iteration, bed_particles, model_particles, x_lim, y_lim, fp_out=None,
     plt.colorbar(p_m,orientation='horizontal',fraction=0.046, pad=0.1,label='Particle Age (iterations since last hop)')
     plt.title(f'Iteration {iteration}')
 
-    if fp_out is None:
+    if out_location is None:
         plt.show()
     else:
         filename = f'iter{iteration}.png'
-        plots_path = fp_out + filename
+        plots_path = out_location + filename
         plt.savefig(plots_path, format='png',)
-        
     return
 
-def stream_gif(start, stop, fp_out, out_name):
-    """Plot gif of the streambed
-    
-    Args: 
-        start: first iterations to plot
-        stop: final iteration to plot
-        fp_out: save location
-        out_name: filename 
-    """
-    in_filename = 'iter{i}.png'
-    out_filename = out_name
-
-    fp_in = dir + in_filename
-    fp_out = dir + out_filename
-
-    step = 1
-    images = []
-
-    for i in range(start, stop, step):
-        im = Image.open(fp_in.format(i=i))
-        images.append(im)
-
-    images[0].save(fp_out, save_all=True, append_images=images[1:], optimize=True, duration=150, loop=0)
-
-def crossing_info(particle_crossing_list, iterations, subsample, fp_out=None, out_name=None):
+def downstream_boundary_hist(particle_crossing_list, iterations, out_location=None, out_name=None):
     """Histogram of downstream particle crossings per iteration
     
     Args:
         particle_crossing_list: array of # of crossings per iteration
         iteration: number of iterations
         subsample: value to subsample by (use if data too large)
-        fp_out: save location
-        out_name: filename (ignored if fp_out not set) 
+        out_location: save location. Default=None will print plot to screen
[…]
