+++ src/tests/test_utils.py
@@ -8,7 +8,7 @@ Todo:
     A headache to resolve, but not impedeing functionality.
 """
 import unittest
-from sbelt import utils
+from ..sbelt import utils
 
 class TestValidateArguments(unittest.TestCase):
 
