### 2022-03-18 de50c80
Create DEFAULT_PARAMS.md
This markdown document describes the default values used by py_SBeLT and why.
