### 2022-03-19 12e6ff3
Update DEFAULT_PARAMS.md
Update default parameters explanation.
