### 2022-03-19 31aaf45
Update THEORY.md
Updated for changes made by eliminating param.yaml
