+++ THEORY.md
@@ -1,85 +1,87 @@
 ## Theory Background
 
-Sediment particle transport by **'py_SBeLT'** is simplified from natural rivers in two key ways. First, the model simulates transport along a profile that is one
-particle wide. Second, the causes of particle motions (e.g. local velocities or shear stresses) are represented in the model through the resulting kinematics--
-namely the number of entrainment events and the particle hop or travel distance. These two quantities alone are the main components that drive transport and the 
-resulting behavior that emerges from model simulations with **'py_SBeLT'**. The remainder of the code is primarily keeping track of particle motions, the changing 
-bed conditions and locations along the bed that are available for deposition. 
+Sediment particle transport by **'py_SBeLT'** is simplified from natural rivers in two key ways. First, the model simulates transport along a profile 
+that is one particle wide. Second, the causes of particle motions (e.g. local velocities or shear stresses) are represented in the model through the 
+resulting kinematics--namely the number of entrainment events and the particle hop or travel distance. These two quantities alone are the main components 
+that drive transport and the resulting behavior that emerges from model simulations with **'py_SBeLT'**. The remainder of the code is primarily keeping 
+track of particle motions, the changing bed conditions and locations along the bed that are available for deposition. 
 
 ### Entrainment
 
-With this general set-up, transport in **'py_SBeLT'** is treated as a stochastic process. The number of entrainment events and the particle travel distance are
-random variables that are indexed to time and space. Entrainment events are sampled from the Poisson probability mass function (pmf) (Fig. 1). The Poisson pmf
-expresses the probability that a specific number of events will occur within a time interval according to the rate constant &#955;. This means particle 
-entrainment is treated as independent events between the **'num_subregions'** (see readme.md and paper.md) and between each iteration. For example, 
-&#955; is specified within the **'parameter.yaml'** file as &#955;<sub>1</sub>. Therefore, the Poisson pmf is fixed for any given simulation. However, for each
-iteration and subregion a new value is randomly sampled from the Poisson pmf to define the entrainment events. 
+With this general set-up, transport in **'py_SBeLT'** is treated as a stochastic process. The number of entrainment events and the particle travel 
+distance are random variables that are indexed to time and space. Entrainment events are sampled from the Poisson probability mass function (pmf) (Fig. 
+1). The Poisson pmf expresses the probability that a specific number of events will occur within a time interval according to the rate constant &#955;. 
+This means particle entrainment is treated as independent events between the **'num_subregions'** (see README.md and paper.md) and between each 
+iteration. Therefore, the Poisson pmf is fixed for any given simulation (see DEFAULT_PARAMS.md). However, for each iteration and subregion a new value is 
+randomly sampled from the Poisson pmf to define the entrainment events. 
 
 |![Image](/figures/poisson.png)
 |:--:| 
 | *Figure 1. Example Poisson pmf used to set the number of entrainment events in **`py_SBeLT`**.* |
 
-Use of the Poisson pmf is linked to sediment transport theory and supproting experimental results which show **'rarefied'** transport (Furbish et al., 2016)
-can be characterized as a Poisson process when entrainment includes effects related to fluid phenomena under steady state transport conditions (Ancey et al., 
-2008). Collective related entrainment effects are not represented (Ancey et al., 2008; Lee and Jerolmack, 2018). We have tested **'py_SBeLT'** for a range of rate 
-constant values (see readme.md). Results from our testing reveal that the value specified for &#955;<sub>1</sub> along with the **'num_subregions'** controls
-the intensity, or magnitude of transport. This makes intuitive sense because the local transport intensity is directly linked to the number of sediment particles 
-entrained from the bed surface.
[…]
