### 2022-03-19 df06dd7
Update paper.md
Updated to remove reference to param.yaml
