+++ paper/paper.md
@@ -56,23 +56,23 @@ flux is represented as a counting phenomenon where the number of particles in mo
 variable [@Ancey:2008]. The model supports ensemble simulations so that repeat numerical experiments can be conducted efficiently,
 or the problem can be probed across a range of input parameter values (discussed below).
 
-`pySBeLT` is run forward in time according to default or user specified parameter values in **'param.yaml'** (see README.md for 
+`pySBeLT` is run forward in time according to default or user specified parameter values (see README.md and DEFAULT_PARAMS.md for 
 more details). After initialization, `pySBeLT` first constructs a bed of fixed particles of set_diam in both the downstream and 
 cross-stream dimensions (one particle wide in the present build), and over a downstream domain length **'x_max'**. Bed surface particles
 of **'set_diam'** are then randomly placed at vertices between fixed bed particles until the **'pack_density'** is met. Vertices are defined 
 by a contact point between two adjacent particles. The bed of surface particles is then separated into **'num_subregions'**, and at this 
-point the forward simulations are ready to commence. The subregion boundaries are located by: (1) the user specifies the **'num_subregions'** in the 
-**'param.yaml'** file (see Readme.md), and (2) the subregion boundaries occur at domain locations set by a distance = **'x_max'** / **'num_subregions'**.
+point the forward simulations are ready to commence. The subregion boundaries are located by: (1) default or user specified **'num_subregions'**(see 
+README.md and DEFAULT_PARAMS.md), and (2) the subregion boundaries occur at domain locations set by a distance = **'x_max'** / **'num_subregions'**.
 
 Simulation iterations involve three steps (Fig. 1): (1) the number of particle entrainment events per **'num_subregions'** are drawn from a Poisson pmf, 
 and this is done randomly for each numerical step up to **'n_iterations'**; (2) surface particles from each subregion are randomly selected 
 for entrainment, and if there are insufficient surface particles available for entrainment, then all available particles are moved; (3) each 
-entrained particle moves a distance according to a randomly sampled value from either the normal or lognormal distribution (see THEORY.md for more details), and 
-is placed at the nearest vertex between two particles that is available for placement. Placed particles are permitted to stack up to the **'level_limit in 
-height'**. Particles are not permitted to travel to the same available vertex. When this occurs the discrepancy is addressed by randomly sampling for one particle 
-to deposit at the identified vertex and the remaining particle(s) are placed at the nearest available vertices. Travel distances of particles that exceed 
-**'x_max'** are returned and queued at the upstream boundary, and are introduced back into the domain at the next numerical step according to travel distance 
-sampling described above. This overall process repeats for the specified **'n_iterations'**.
+entrained particle moves a distance according to a randomly sampled value from either the normal or lognormal distribution (see THEORY.md for more 
+details), and is placed at the nearest vertex between two particles that is available for placement. Placed particles are permitted to stack up to the 
+**'level_limit in height'**. Particles are not permitted to travel to the same available vertex. When this occurs the discrepancy is addressed by 
+randomly sampling for one particle to deposit at the identified vertex and the remaining particle(s) are placed at the nearest available vertices. Travel 
+distances of particles that exceed **'x_max'** are returned and queued at the upstream boundary, and are introduced back into the domain at the next 
+numerical step according to travel distance sampling described above. This overall process repeats for the specified **'n_iterations'**.
 
 `pySBeLT` tracks a number of different parameters through a simulation: the vertical and horizontal positions of every particle center, 
 the randomly sampled number of entrainment events, the number of particles actually entrained, the randomly sampled particle travel 
