### 2022-03-20 fa19cad
Update README, plotting docsstring/parameters, docs/ structure
