+++ README.md
@@ -6,139 +6,34 @@ Rivers transport sediment particles. Individual particles can exhibit transport 
 
 ## Installation
 
-First, [clone the Py_SBeLT repository](https://docs.github.com/en/repositories/creating-and-managing-repositories/cloning-a-repository) in a desired location.
-
-Then install the required Python dependancies:
+### Quick Installation
 
 ```bash
- pip3 install -r requirements.txt
+pip install sbelt
 ```
 
-## Nomeculture
+### Installation from Source
 
-- **Stream**: An 2D (x,y) area/domain consisting of bed particles and model particles. The length of the stream is set via parameters while the height has no explicit value. The left-most point of the stream is considered the upstram boundary while the right-most is considered the downstream boundary.
-- **Bed Particle**: A fixed particle comprising the stream bed. At no point in the model simulation do bed particles move. Bed particles represent the lowest elevation in the stream.
-- **Model Particle**: A particle available for an entrainment event. All model particles sit on top of the fixed bed particles. Model particles move and have their location and other metrics tracked per-iteration for the duration of the model simulation.
-- **Supporting Particle**: Particles that hold up other particles (i.e in a stack of 3 particles, the bottom two particles are supporting the top). Both bed and model particles can be considered supporting particles at a given point in the model simulation.
-- **Event Particle**: A particle which has been selected to undergo an entrainment event. Only model particles can be considered event particles.
-- **Available Vertex**: A horizontal location in the stream where a model particle is permitted to rest/land based on the selected particle travel or hop distance.
-- **Desired Hop**: The specified travel or hop distance of an entrained particle based on random sampling informed by parameters.
-
-## Parameters
-
-The model reads in parameters from a yaml file. The default parameters file for the model is `param.yaml`, located in the **`model/parameters/`** directory:
-
-```
-project
-│   README.md
-│   requirements.txt
-└───model
-|   └───parameters
-│   │   |   param.yaml
-|   |   |   schema.yaml
-│   │   run.py
-│   │   ...
-```
-
-These parameters can be changed for whatever type of run you desire! However, all parameters have data type requirements and some have minimum and maximum permitted values. The parameters are validated at the start of the run and any incompatible entries will cause an error message to be raised. The following is a table of the parameters:
-
-| Parameter | Type | Tested Range | Description |
-| ----------- | ------- | ------ | ----------- |
-| pack_density | float | 0.50, 0.80 | The packing fraction of the model particles (-) |
-| x_max | int | 100, 1000 | Length of the domain in the streamwise direction (mm) |
-| set_diam | int or float =+ 0.5 | 0.5, 10 | Grain diameter (mm) |
-| num_subregions | int | 1, 10 | The number of bed subregions |
-| level_limit | int | 1, 3 | The maximum number of levels permitted (i.e how many particles high to stack)  |
-| n_iterations | int | 1, 1000000 |The number of iterations to run |
-| lambda_1 | int | 1, 5 | Lamba for poisson dist., used to determine the number of entrainment events |
-| normal_dist | boolean | - | Flag for which distribution to sample from for hop calculations. True=Normal, False=logNormal |
-| mu | float | 0.25, 2 | Mean/expectation of the logNormal/Normal distribution for hop calculations |
-| sigma | float | 0.25, 0.50 | Standard deviation of logNormal/Normal distribution for hop calculations|
-| data_save_interval | int | 1, 2 | How often to record model particle arrays (e.g 1=save every iteration, 2=save every other) |
[…]
