### 2022-03-20 1d341a0
Update setup for new PyPi dist, clear notebook output
