### 2022-03-21 989ec45
Update README.md
Added some attribution placeholders as well as acknowledgement
