### 2022-03-21 357c110
Final changes. Particle array diagram + loop age --> loop count
