### 2022-03-28 #4091 @kbarnhart
@szwiep and @smchartrand thanks for these updates, they are great.

I have a couple of very minor comments and I've proposed changes that I think would be beneficial in a pull request. Most of the recommendations are to provide more links to other files in the Readme and to structure the documentation a bit. Feel free to modify any of my recommendations, they are just suggestions. The only thing I didn't include was actual text describing how a user might test an installation. Forgive me if that was somewhere I didn't notice.

I will now start looking for reviewers.
