### 2022-03-28 #4091 @kbarnhart
:wave: @pfeiffea @tdoane would you be willing to review the submission "py_SBeLT: A Python software for stochastic sediment transport under rarefied conditions" for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html. I think your backgrounds in sediment transport would make you valuable reviewers of this contribution.

If you are interested and able, please read our [conflict of interest](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) before agreeing.

We would ask for reviews within 4–6 weeks, but that generally means starting the process sooner since it tends to be iterative between the reviewers and author(s). You can read more about the process here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

Thanks for your consideration and let me know if you have any questions (here or at krbarnhart@usgs.gov)
