### 2022-03-31 7f3b96b
Update paper.md
Update paper for version 1.0.1 changes.
