### 2022-04-05 #4282 @szwiep
Hi all, I forgot to include the updated `paper.md` in the pre-review revisions. I realized my mistake and just pushed it to the main branch now. Hopefully this doesn't cause any difficulties. @kbarnhart should I re-generate the pdf?
