### 2022-04-22 #4282 @pfeiffea
First off, well done @szwiep (and @smchartrand). This is a very well organized, well documented bit of software. It is a simple and elegant little tool. My edits are relatively minor. I have a few lingering “to-dos” on my review checklist, which I’ll deal with in the coming days.

# Basic Usage Notebook:
This is an excellent notebook. It is not verbose, but gives me a great sense for how to use and interpret results of pySBeLT.
Minor suggestion: It would be great to add the default parameters to the “parameters and running” table.

# Data Storage Notebook:
Accessing Final Metrics section you’re missing a “.hdf5” at the end of the file name in the 2nd and 3rd to last code blocks. Make sure to restart the kernel and clear all outputs (or run all) in your final submission. The version up on your repo currently shows that you’ve re-run cells several times without restarting.

# Manuscript:
Ln 12: I’d suggest editing to “… a function of local water flow conditions and the grain size distribution of the bed material.”
Ln 12-14: The sentence that starts “A predictive approach…” seems like your opportunity to succinctly state the gap pySBeELT is trying to fill. But, as written, it feels a bit muddled (“the approach introduces challenges”). Do you mean something like “This standard approach to modeling bedload transport obscures(? Neglects?) the mechanistic reality of sediment transport: the bed material flux is the integrated transport of individual stochastic particle motions.”
Ln 17: Instead of “examine connections”, how about “reveal the relationship” or “illustrate the relationship” (the user will be the one examining, the tool will make the examination possible)?
Ln 38: “motivated by a..” suggests that your work was motivated by some specific earlier published work. If so, cite. Or maybe “pySBeLT takes a birth-death….. approach to modeling bedload transport”.
Ln 46: set_diam should have quotes around it, right? ‘set_diam’
Ln 51-54: I think these last two sentences should be combined or re-ordered.
Figure 1. Add a flow direction arrow.
Figure 2. The lower panel is too small to read. Add the units of particle flux to the y axis of the upper plot.

Fix citation issues:
? in Wainwright 2014 and Wu 2019
Some first names or initials included in the in-text citations.

# THEORY.md
Entrainment section, second paragraph has some odd number typos interspersed in the text. Take a look.
@kbarnhart should this section be a part of the paper? It seems to me that it describes both the rationale and function of the model, so maybe? Or is it better to keep separate?
