### 2022-04-26 #4282 @kbarnhart
@pfeiffea many thanks for your review! 🎉

Regarding your question:
> @kbarnhart should this section be a part of the paper? It seems to me that it describes both the rationale and function of the model, so maybe? Or is it better to keep separate?

I tend to recommend that theory lives with the main documentation (here, the file theory.md) rather than in the paper because the theory is then more closely tied with what a typical user is looking at when they are learning to use the software. In contrast, the paper would contain the type of information that a potential user might use to determine if this software could solve their problem.

I agree with you that in this case, the theory.md file includes both rationale and function (e.g., discussion of results of model testing). A more traditional theory document would probably just have the rationale and mathematical description. A second document might discuss how the theory interacts with the implementation (e.g., the discussion of how discretization interacts with entrainment "This means particle entrainment is treated as independent events between the 'num_subregions' (see README.md and paper.md) and between each iteration.")

If either you (@pfeiffea) or @tdoane found that it was hard to "see" the theory as described by "theory.md", then it may be worth recommending to @szwiep and co-authors to revise. If not, I think that it is fine as it is.
