### 2022-04-27 #4282 @tdoane
Hi @szwiep, @smchartrand, and @kbarnhart,

Here's my relatively minor comments. I have not used GitHub much in a collaborative way, so please let me know if I should post these differently!

This is an impressive piece of work that will provide a useful tool for teaching and research related to stochastic sediment transport. We can so simply state that sediment transport is the result of distributions of particle travel distance and particle rest times. However, when it comes to modeling this behavior we quickly run in to many complications related to keeping track of particle positions, burial, and particle history. This is a very clean treatment of the problem and the authors have clearly thought deeply about it. I have relatively minor comments that mostly relate to the paper. I want to emphasize that I think this is an excellent tool, and in particular that there is potential to use this to effectively illustrate the value of probabilistic approaches and the cosnequences of different distributions.

Paper:
Why pySBeLt? It is an acronym?

Statement of need, 1st paragraph: roughly twice the shear stress threshold for motion? I assume?

Statement of need, 5th paragraph: For the periodic boundary condition: If a particle passes the lower boundary, it is placed at the upstream boundary, but does not continue with the motion that passed it over the downstream boundary, right? So the travel distance that sends it over the boundary is not recorded except that it exceeds a certain distance?

Statement of need: What is the significance of the length of the flume to the distribution of particle travel distances. Given the quasi periodic boundary conditions, this might warrant a brief discussion.

Statment of need: The authors end with a statement about the utility in classrooms. I completely agree and wonder if it could/should be stated near the top.

Theory:

Theory: a normal distribution will have negative travel distances. are these allowed? or is the distribution forced so that particle travel distances are all down stream?

Theory: What about the possibility of adding more probability distributions? Insofar as this has potential to be an excellent teaching tool, do the authors think that there is value in providing more distributions? In particular, a Pareto distribution has a minimum value that may address the issue of isolated piles forming for small or zero mode and it has a heavy tail where the variance becomes undefined for certain shape parameters. This might be an opportunity to illustrate the consequences of heavy-tailed distributions and the influence on the time-series of sediment flux.

Theory, 4th paragraph: "posiitons"

Code:
The only issue was the same one raised by @ampfeif
