+++ paper/paper.bib
@@ -5,7 +5,7 @@
 	doi = {10.1103/PhysRevE.74.011302},
 	number = {1},
 	journal = {Physical Review E},
-	author = {Ancey, Christophe and Böhm, Tobias and Jodeau, Magali and Frey, Philippe},
+	author = {Ancey, C. and Böhm, T. and Jodeau, M. and Frey, P.},
 	month = jul,
 	year = {2006},
 	note = {Publisher: American Physical Society},
@@ -17,7 +17,7 @@
 	volume = {595},
 	doi = {10.1017/S0022112007008774},
 	journal = {Journal of Fluid Mechanics},
-	author = {Ancey, C. and Davison, A. C. and Böhm, T and Jodeau, M and Frey, P},
+	author = {Ancey, C. and Davison, A.C. and Böhm, T and Jodeau, M. and Frey, P.},
 	year = {2008},
 	note = {Publisher: Cambridge University Press},
 	pages = {83--114}
@@ -30,7 +30,7 @@
 	doi = {10.1029/2009JF001260},
 	number = {F2},
 	journal = {Journal of Geophysical Research: Earth Surface},
-	author = {Ancey, Christophe},
+	author = {Ancey, C.},
 	month = jun,
 	year = {2010},
 	keywords = {sediment transport, Exner equation, stochastic differential equation}
@@ -44,7 +44,7 @@
 	doi = {10.1080/00221686.2019.1702594},
 	number = {1},
 	journal = {Journal of Hydraulic Research},
-	author = {Ancey, Christophe},
+	author = {Ancey, C.},
 	month = jan,
 	year = {2020},
 	note = {Publisher: Taylor \& Francis},
@@ -57,7 +57,7 @@
 	url={https://open.library.ubc.ca/collections/ubctheses/24/items/1.0349138}, 
 	doi={10.14288/1.0349138}, 
 	school={University of British Columbia}, 
-	author={Chartrand, Shawn M.}, 
+	author={Chartrand, S.M.}, 
 	year={2017}, 
 	collection={Electronic Theses and Dissertations (ETDs) 2008+}
 }
@@ -67,7 +67,7 @@
   	year      = {2013},
   	publisher = {O'Reilly},
   	title     = {Python and HDF5},
-  	author    = {{Collette}, Andrew}
+  	author    = {Collette, A.}
 }
 
 @phdthesis{Einstein:1937,
@@ -84,7 +84,7 @@
 	doi = {10.1002/2015JF003552},
 	number = {11},
 	journal = {Journal of Geophysical Research: Earth Surface},
-	author = {Fathel, Siobhan L. and Furbish, David Jon and Schmeeckle, Mark W.},
+	author = {Fathel, S.L. and Furbish, D.J. and Schmeeckle, M.W.},
 	month = nov,
 	year = {2015},
 	pages = {2298--2317}
@@ -97,7 +97,7 @@
 	doi = {10.1029/2012JF002352},
 	number = {F3},
 	journal = {Journal of Geophysical Research: Earth Surface},
-	author = {Furbish, David Jon and Haff, Peter K and Roseberry, John C and Schmeeckle, Mark W},
+	author = {Furbish, D.J. and Haff, P.K. and Roseberry, J.C. and Schmeeckle, M.W.},
 	month = sep,
 	year = {2012},
 	keywords = {bed load sediment, Fokker-Planck equation, 1825 Geomorphology: fluvial, 1862 Sediment transport, advection, diffusion},
@@ -111,7 +111,7 @@
 	doi = {10.1002/2016JF003833},
 	number = {7},
 	journal = {Journal of Geophysical Research: Earth Surface},
[…]
