+++ paper/paper.bib
@@ -167,7 +167,7 @@
 }
   
 @article{Wainwright:2015,
-	author = {Wainwright, J. and Parsons, A.J. and Cooper, James R. and Gao, P. and Gillies, J.A. and Mao, L. and Orford, J.D. and Knight, P.G.},
+	author = {Wainwright, J. and Parsons, A.J. and Cooper, J.R. and Gao, P. and Gillies, J.A. and Mao, L. and Orford, J.D. and Knight, P.G.},
 	title = {The concept of transport capacity in geomorphology},
 	journal = {Reviews of Geophysics},
 	volume = {53},
