### 2022-05-09 deb8568
Update paper.md
Address peer review comments by @pfeiffea.
