+++ paper/paper.md
@@ -27,11 +27,13 @@ bibliography: paper.bib
 Granular sediment of various sizes moves downstream along river beds when water flow is capable of entraining particles from 
 the bed surface. This process is known as bed load sediment transport because the particles travel close to the boundary. 
 It is common to treat the transport process as a predictive problem in which the mean transport rate past a stationary 
-observation point is a function of local water flow conditions [@Parker:2008; @Wainwright:2014; @Ancey:2020]. A predictive approach 
-introduces challenges to understanding bed load transport, however, because the stochastic nature of transport due to the 
-movements of individual particles is neglected [@Einstein:1937; @FurbDoane:2021]. Here, we present an open-source Python model, `pySBeLT`, 
-which simulates the kinematics of rarefied particle transport (low rates) as a stochastic process along a riverbed profile. 
-The primary aim of `pySBeLT` is to examine connections between individual particle motions and local transport rates, or the flux.
+observation point is a function of local water flow conditions and the grain size distribution of the bed material 
+[@Parker:2008; @Wainwright:2014; @Ancey:2020]. However, a predictive approach to the bed load problem neglects the stochastic 
+nature of transport due to the movements of individual particles [@Einstein:1937; @FurbDoane:2021], and interactions between
+moving particles and those on the bed surface [@Ancey:2006; @Ancey:2008;LeeJerol:2018]. Here, we present an open-source Python model,
+`pySBeLT`, which simulates the kinematics of rarefied particle transport (low rates) as a stochastic process along a riverbed profile.
+The primary aim of `pySBeLT` is to offer an efficient and reasonable numerical means to help in examining connections between individual
+particle motions and local transport rates, or the flux.
 
 # Statement of need
 
@@ -50,41 +52,44 @@ Because particle motions are controlled by fluid turbulence, the irregular bed s
 [@Ancey:2006; @Ancey:2008; @LeeJerol:2018], the connection between particle movements and the bedload 
 transport rate has been difficult to formulate mathematically. `pySBeLT` provides an extensible framework within 
 Python to numerically examine correlations between upstream particle entrainment rates and travel distances, with downstream 
-flux. `pySBeLT` was motivated by a birth-death, immigration-emigration Markov model for bedload transport. Here, 
-the movements of individual particles are represented by stochastic entrainment, motion, and deposition processes, and sediment 
+flux. `pySBeLT` was motivated by a birth-death, immigration-emigration Markov model for bedload transport [@Ancey:2008; @Ancey:2010]. 
+Here, the movements of individual particles are represented by stochastic entrainment, motion, and deposition processes, and sediment 
 flux is represented as a counting phenomenon where the number of particles in motion above the bed surface is a random 
 variable [@Ancey:2008]. The model supports ensemble simulations so that repeat numerical experiments can be conducted efficiently,
 or the problem can be probed across a range of input parameter values (discussed below).
 
-`pySBeLT` is run forward in time according to default or user specified parameter values. After initialization, `pySBeLT` first constructs a bed of fixed particles of set_diam in both the downstream and 
-cross-stream dimensions (one particle wide in the present build), and over a downstream domain length **'bed_length'**. Bed surface particles
-of **'particle_diam'** are then randomly placed at vertices between fixed bed particles until the **'particle_pack_frac'** is met. Vertices are defined 
-by a contact point between two adjacent particles. The bed of surface particles is then separated into **'num_subregions'**, and at this 
-point the forward simulations are ready to commence. The subregion boundaries are located by: (1) the user specifies the **'num_subregions'**, and (2) the subregion boundaries occur at domain locations set by a distance = **'bed_length'** / **'num_subregions'**.
+`pySBeLT` is run forward in time according to default or user specified parameter values. After initialization, `pySBeLT` first constructs
[…]
