### 2022-05-09 e095c2d
Update paper.md
Additional edits based on peer review of @tdoane
