### 2022-05-09 de4ea28
Update THEORY.md
Edits to address peer review by @tdoane
