+++ peer_review/PRreadme.md
@@ -1,1 +1,1 @@
-This folder contains a response to comments document for the peer reviews completed for the JOSS submission.
+This folder contains a response to comments document for the [peer reviews completed for the JOSS submission](https://github.com/openjournals/joss-reviews/issues/4282).
