### 2022-05-12 #4282 @szwiep
Hi @kbarnhart, @smchartrand and I have completed the suggested changes. Shawn has also prepared a short [response to comments document ](https://github.com/szwiep/py_SBeLT/blob/master/peer_review/Zwiep%20and%20Chartrand%20JOSS%20Response%20to%20Comments%20May%202022.pdf) that can hopefully be some help during the re-review process!
