### 2022-05-13 szwiep/py_SBeLT#22 @kbarnhart
Small number of post review comments
@szwiep and @smchartrand

Two small comments from me before I ask the reviewers to finalize their reviews.

1. In the paper and the readme, I'd recommend defining the acronym pySBeLT. Stochastic BedLoad Transport? Something else?
2. There are a few citations that are misformatted (lines 12, 15, and 29,  Wainwright:2014, Wu:2019; LeeJerol:2018)

https://github.com/openjournals/joss-reviews/issues/4282
