### 2022-05-13 cd200d5
Update README.md
Address acronym comment
