### 2022-05-13 1243613
Update paper.md
Address acronym comment
