### 2022-05-13 b5d98ab
Update paper.md
Address citation issues
