### 2022-05-13 szwiep/py_SBeLT#22 @smchartrand
@kbarnhart - thanks for catching these. I have addressed the issues. Apologies and have a nice weekend.
