### 2022-05-24 #4282 @smchartrand
@kbarnhart @tdoane I have edited the sentence in the Theory.md which @tdoane discusses in the post today. I simply removed the clause pertaining to the lognormal distribution - the revised sentence reads: The Pareto distribution also has the advantage of
being heavy-tailed where the variance becomes undefined for certain shape parameters.

thanks.
