### 2022-05-24 #4282 @pfeiffea
Thanks for the prod @kbarnhart.

I'm happy with the changes the authors have made addressing my comments.

The one final checkbox on my review list was testing. I think this was my programming naivete, but simply following the *Testing your installation* code didn't work for me:
`python -m unittest src/tests/test_*` gave an import error. Should that wildcard '*' work as written? I'm running it on Windows 10 in anaconda powershell prompt, in case that matters. Running `python -m unittest src/tests/test_utils.py` and `python -m unittest src/tests/test_logic.py` ran for me, though.

The latter, however, gives a bunch of errors: `ERROR:root:No left supporting particle at 0.75for particle 0.0`, all for particle 0.0.

@smchartrand and @szwiep I suspect there's something special with particle 0.0..? Perhaps you just need an exception for that error?
