### 2022-05-25 #4282 @szwiep
Hi @pfeiffea! I'll look into the issue with the `python -m unittest src/tests/test_*` call tonight. I'm hoping it's just an incompatibility with Windows. If you have a chance, could you post the import errors you're getting?

As for the `particle 0.0` business, those are forced errors in the `TestFindSupports.test_no_supports_available_returns_value_error` unit test. It tests that an error is thrown if a model particle doesn't have 2 supporting particles. So in this case, the errors being thrown are a good sign. If you think it would be more clear, I could play around  with the the output configuration to disable errors messages during unit testing?
Either way I'll update the `testing` section to more clearly describe the output one should expect. Thank-you for the input!
