### 2022-05-25 ada8143
Update .circleci test call
