### 2022-05-25 d397469
Update README.md
Update README to match related updated in pull request #23
