### 2022-05-25 #4282 @szwiep
Hi again @pfeiffea, I believe I've addressed your comments in pull requests #23 and #24 in the project repository. The new unittest commands appear to be working on both Linux and Windows systems and the stderr from passing tests have been suppressed to help make the results more clear to users. Let me know if these changes are appropriate. Thanks again!
