### 2022-05-31 44386f8
Update paper.md
Add missing @ in reference.
