### 2022-05-31 #4282 @kbarnhart
@szwiep @smchartrand, please merge https://github.com/szwiep/py_SBeLT/pull/25 and then re run the command

`@editorialbot generate pdf`

here.

Verify that the Lee and Jerolmack citation formats correctly (it was just missing an @). Presuming this resolves that issue, then do the following:

- [ ] Make a tagged release of your software.
- [ ] List the DOI of the archived version here.
- [ ] List the version tag number of the archived version here.

I will then update the paper's metadata and recommend it be accepted. At that point the handling editor in chief will manage final steps.
