### 2022-06-02 #4282 @smchartrand
@kbarnhart - here you go, and thanks again for handling the work and your support through the process - it is much appreciated.

1. Make a tagged release of your software. - **Completed by @szwiep today at 10:59**

2. Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository) - **Completed today. Made archive at figshare.**

3. Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID. - **Completed today. Metadata are the same, including orcid id's.**

4. List the DOI of the archived version here. - **https://doi.org/10.6084/m9.figshare.19967552.v2**

5. List the version tag number of the archived version here. - **v1.0.2**
