### 2022-06-03 #4282 @kbarnhart
@szwiep and @smchartrand -- I've recommended that your submission be accepted and published. One of the JOSS Editors in Chief (ping @openjournals/joss-eics) will handle the submission from this point.

Many thanks to @pfeiffea and @tdoane for your reviews.

Thanks to all for your contributions to JOSS.
