### 2022-06-03 #4282 @smchartrand
Thanks very much to you all - @kbarnhart @danielskatz @tdoane @pfeiffea.

It has been a real pleasure working through the review process for JOSS.

Have a wonderful weekend.
