### 2022-06-03 #4282 @danielskatz
Once these two issues are fixed, the paper looks ready to go to me
