### 2022-06-03 #4282 @smchartrand
@danielskatz

For the title, how about: py_SBeLT: Python **model** for stochastic sediment transport under rarefied conditions?

I am glad you raised this issues. I am hesitant to use "package", and now "software" since it is really a numerical model. Thoughts?
