### 2022-06-03 #4282 @danielskatz
"A Python model" sounds better grammatically to me, though I think the model is implemented in software, so I could see reasons to state this is software too, particularly as a JOSS paper.
