### 2022-06-03 9318b37
Update paper.md
Edits per EiC comments at JOSS
