### 2022-06-03 #4282 @smchartrand
@danielskatz

Updated DOI for figshare archive: https://doi.org/10.6084/m9.figshare.19967552.v3
